Added an interface class called Shape, which introduces the strategy pattern. This adds the ability to decouple the algorithm from the object using the algorithm. Square no longer has to extend from rectangle and overwrite its methods, rather it implements from the Shape interface and declares how it wants to use the methods
