+++ src/AccessoryDecorator.java
@@ -1,6 +1,5 @@
 public abstract class AccessoryDecorator extends Rectangle {
 
 
-
   }
