+++ src/AccessoryDecorator_structure.puml
@@ -1,0 +1,4 @@
+@startuml
+
+
+@enduml
