+++ src/Shape.java
@@ -1,0 +1,9 @@
+import java.awt.*;
+
+public interface Shape {
+
+
+        void setLocation(int x, int y);
+        void draw(Graphics g);
+
+}
