+++ src/SuperSuperJump.java
@@ -7,13 +7,16 @@ public class SuperSuperJump extends AccessoryDecorator {
         this.r=r;
     }
 
-
-
     public Color addColour() {
         return r.addColour();
     }
 
     public int addAccessory() {
         return 85;
+    }
+
+    @Override
+    public void draw(Graphics g) {
+
     }
 }
